requirements.txt
discord.py==2.2.3
openai==0.27.8
python-dotenv==1.0.0
pynacl==1.5.0

